requirements.txt
# requirements.txt (UPDATED)
langchain==0.2.1
langgraph==0.0.63
langchain-openai==0.1.7
langchain-community==0.2.1
chromadb==0.5.0
pypdf==4.2.0
tavily-python==0.3.3
python-dotenv==1.0.1
sentencepiece==0.2.0

# Pinned versions to ensure stability
transformers==4.41.2
datasets==2.19.2
accelerate==0.30.1
torch==2.3.1